Move clean_server_state fixtures to integration/.
