+++ robot-server/tests/conftest.py
@@ -1,4 +1,3 @@
-import asyncio
 import json
 import os
 import pathlib
@@ -6,7 +5,7 @@ import tempfile
 from datetime import datetime, timezone
 from mock import MagicMock
 from pathlib import Path
-from typing import Any, Callable, Generator, Iterator, cast
+from typing import Callable, Generator, Iterator, cast
 from typing_extensions import NoReturn
 from decoy import Decoy
 
@@ -35,7 +34,6 @@ from robot_server.hardware import get_hardware
 from robot_server.versioning import API_VERSION_HEADER, LATEST_API_VERSION_HEADER_VALUE
 from robot_server.service.session.manager import SessionManager
 from robot_server.persistence import get_sql_engine, create_sql_engine
-from .integration.robot_client import RobotClient
 from robot_server.health.router import ComponentVersions, get_versions
 
 test_router = routing.APIRouter()
@@ -169,57 +167,6 @@ def server_temp_directory() -> Iterator[str]:
     os.environ["OT_API_CONFIG_DIR"] = new_dir
     config.reload()
     yield new_dir
-
-
-@pytest.fixture()
-def clean_server_state() -> Iterator[None]:
-    # async fn that does the things below
-    # make a robot client
-    # delete protocols
-    async def _clean_server_state() -> None:
-        port = "31950"
-        async with RobotClient.make(
-            host="http://localhost", port=port, version="*"
-        ) as robot_client:
-            await _delete_all_runs(robot_client)
-            await _delete_all_protocols(robot_client)
-
-    yield
-    asyncio.run(_clean_server_state())
-
-
-# TODO(jbl 2023-05-01) merge this with ot3_run_server, along with clean_server_state and run_server
-@pytest.fixture()
-def ot3_clean_server_state() -> Iterator[None]:
-    # async fn that does the things below
-    # make a robot client
-    # delete protocols
-    async def _clean_server_state() -> None:
-        port = "31960"
-        async with RobotClient.make(
-            host="http://localhost", port=port, version="*"
-        ) as robot_client:
-            await _delete_all_runs(robot_client)
-            await _delete_all_protocols(robot_client)
-
-    yield
-    asyncio.run(_clean_server_state())
-
-
-async def _delete_all_runs(robot_client: RobotClient) -> None:
-    """Delete all runs on the robot server."""
-    response = await robot_client.get_runs()
-    run_ids = [r["id"] for r in response.json()["data"]]
-    for run_id in run_ids:
-        await robot_client.delete_run(run_id)
-
-
-async def _delete_all_protocols(robot_client: RobotClient) -> None:
-    """Delete all protocols on the robot server"""
-    response = await robot_client.get_protocols()
-    protocol_ids = [p["id"] for p in response.json()["data"]]
-    for protocol_id in protocol_ids:
-        await robot_client.delete_protocol(protocol_id)
 
 
 @pytest.fixture
